Fix 40 bit locs after rebase

+++ clang/parse/term/nam.c
@@ -1,7 +1,7 @@
-fn Term parse_term(PState *s, u16 depth);
+fn Term parse_term(PState *s, u32 depth);
 
 // ^name or ^(f x)
-fn Term parse_term_nam(PState *s, u16 depth) {
+fn Term parse_term_nam(PState *s, u32 depth) {
   parse_skip(s);
   if (parse_peek(s) == '(') {
     // ^(f x) -> DRY(f, x)
@@ -12,7 +12,7 @@ fn Term parse_term_nam(PState *s, u16 depth) {
     return term_new_dry(f, x);
   } else {
     // ^name -> NAM(name)
-    u16 nam = parse_name_num(s);
+    u32 nam = parse_name_num(s);
     return term_new(0, NAM, nam, 0);
   }
 }
